Refactor common test code into test_util.ts
