+++ api-service/webapp-demo/src/app/identity_job.component.css
@@ -1,4 +1,4 @@
-.demo-tab-content {
+.demo-content {
   padding: 16px;
 }
